[NETBEANSINFRA-189]: Fix regression introduced by [NETBEANSINFRA-188]
The optimization  - implemented in #8 - introduced a serious regression
by truncating the whole jar file instead of modifying a single entry in
the jar file.

This PR fixes this by applying another approach. Using the in Java 7
introduced Files API, one can create a virtual filesystem of a zip file
and read or write single entries.

In my tests the now modfied jar files are no longer truncated
and the whole module suite can be compiled and executed successfully.
The intended speedup retains.