[NETBEANSINFRA-264] Include runtime module libraries
The nbm Maven plugin packs module libraries together with the module
and adds them to the "Class-Path" entry of the manifest. These
libraries are stored at netbeans\modules\ext within the module's
nbm file. However, only libraries with scope "compile" are recognized.
If the module declares dependencies (direct or transitive) with scope
"runtime" in its pom file, these libraries won't be included in the
module's nbm file nor will they be added to the classpath.

Add the parameter includeRuntimeModuleLibraries to
NetBeansManifestUpdateMojo to choose between the existing behaviour,
i.e. adding module libraries with scope "compile" only. Or changing
the creation of the dependency tree so that module libraries with
scope "compile" as well as "runtime" will be recognized.