Move off from Maven2 codebase (maven-compat)
Refactored codebase to make it "modern" Maven 3
plugin and stop using Maven 2 APIs (maven-compat).

At the same time, move off legacy Plexus XML for
components, and define ArtifactHandlers and Lifecycles
in source.

Post merge TODO:
Once in future, the method Artifacts#pathOf should
go away, but for now is kept tightly isolated, but
the usage of it should not spread that currently is
used.